+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 + Have exceptions in the background recv thread tasks bubble up as inner exceptions to preserve the stack trace for better debugging
 + Fix authentication with explicit credential on Windows
 + Added `-TracePath` to `New-OpenADSessionOption` to help debug raw LDAP traffice exchanged in a session.
++ Fix credential prompt when specifying `-Credential my-username` for a `PSCredential` parameter
 
 ## v0.1.0-preview3 - 2022-03-22