+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 + Fix authentication with explicit credential on Windows
 + Added `-TracePath` to `New-OpenADSessionOption` to help debug raw LDAP traffice exchanged in a session.
 + Fix credential prompt when specifying `-Credential my-username` for a `PSCredential` parameter
++ Have `Get-OpenADWhoami` return an object with more details on the LDAP session, like the domain controller DNS name, URI, and authentication method used.
+  + The returned username value will also strip the leading `u:` prefix if it is present
++ Added the `DomainController` property to the `OpenADSession` class to help identify the domain controller the session is connected to
 
 ## v0.1.0-preview3 - 2022-03-22